Reused logging and timing of getting mapping
