+++ sciety_labs/providers/semantic_scholar_bigquery_mapping.py
@@ -1,5 +1,4 @@
 import logging
-from time import monotonic
 from typing import Any, Mapping, Sequence
 
 import pyarrow.compute as pc
@@ -27,24 +26,15 @@ class SemanticScholarBigQueryMappingProvider(
             **kwargs
         )
 
-    def get_semantic_scholar_paper_ids_by_article_dois_map(
+    def do_get_semantic_scholar_paper_ids_by_article_dois_map(
         self,
         article_dois: Sequence[str]
     ) -> Mapping[str, str]:
         arrow_table = self.get_arrow_table()
-        start_time = monotonic()
         filtered_arrow_table = arrow_table.filter(
             pc.field('article_doi').isin(article_dois)
         )
-        paper_ids_by_article_dois_map = dict(zip(
+        return dict(zip(
             _to_str_list(filtered_arrow_table['article_doi']),
             _to_str_list(filtered_arrow_table['paper_id'])
         ))
-        end_time = monotonic()
-        LOGGER.info(
-            'Looked up paper ids, article_dois=%r, map=%r, time=%.3f seconds',
-            article_dois,
-            paper_ids_by_article_dois_map,
-            (end_time - start_time)
-        )
-        return paper_ids_by_article_dois_map
